web_service/requirements.txt
authlib
fastapi
uvicorn[standard]
pillow
torch
torchvision
numpy
opencv-python
scikit-learn
python-multipart
PyJWT
pydantic
boto3
python-dotenv
PyJWT                
pydantic 
torch              
Pillow       
boto3
authlib
werkzeug
flask
requests
httpx
pymemcache